# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReportSummary {
  id: string;
  slug: string;
  title: string;
  industry?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export type ReportRowInputs = { [key: string]: unknown };

export type ReportRowOutput = { [key: string]: unknown };

export interface ReportRow {
  id: string;
  slug: string;
  userId: string;
  title: string;
  industry?: string | null;
  inputs: ReportRowInputs;
  output: ReportRowOutput;
  status: string;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ReportCreated {
  id: string;
  slug: string;
}

export type CreateReportBodyInputs = { [key: string]: unknown };

export type CreateReportBodyOutput = { [key: string]: unknown };

export interface CreateReportBody {
  title: string;
  industry?: string | null;
  inputs: CreateReportBodyInputs;
  output: CreateReportBodyOutput;
}

export interface UpdateReportStatusBody {
  status: string;
  note?: string | null;
}

export interface NotificationItem {
  id: string;
  kind: string;
  title: string;
  body?: string | null;
  url?: string | null;
  readAt?: string | null;
  createdAt: string;
}

export interface CommentItem {
  id: string;
  reportId: string;
  userId: string;
  section?: string | null;
  body: string;
  createdAt: string;
}

export type CommentsResponseProfiles = { [key: string]: unknown };

export interface CommentsResponse {
  comments: CommentItem[];
  profiles: CommentsResponseProfiles;
}

export interface AddCommentBody {
  body: string;
  section?: string | null;
}

export interface ProfileItem {
  userId: string;
  displayName?: string | null;
  avatarUrl?: string | null;
}

export interface AutofillBriefBody {
  brief: string;
}

export type CompleteFieldBodyInputs = { [key: string]: unknown };

export interface CompleteFieldBody {
  field: string;
  partial: string;
  inputs: CompleteFieldBodyInputs;
}

export type AnalyzeConceptBodyInputs = { [key: string]: unknown };

export interface AnalyzeConceptBody {
  inputs: AnalyzeConceptBodyInputs;
}

export type GetReportOutput200 = { [key: string]: unknown };

export type ListCommentsParams = {
  section?: string;
};

export type AutofillBrief200 = { [key: string]: unknown };

export type CompleteField200 = { [key: string]: unknown };

export type AnalyzeConcept200 = { [key: string]: unknown };