# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /comments/{reportId}:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Comments with profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsResponse"
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
...
  /ai/autofill-brief:
    post:
      operationId: autofillBrief
      tags: [ai]
      summary: Auto-fill form from a brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutofillBriefBody"
      responses:
        "200":
          description: Draft inputs
          content:
            application/json:
              schema:
                type: object
  /ai/complete-field:
    post:
      operationId: completeField
      tags: [ai]
      summary: AI complete a field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteFieldBody"
      responses:
        "200":
          description: Completed text
          content:
            application/json:
              schema:
                type: object
  /ai/analyze-concept:
    post:
      operationId: analyzeConcept
      tags: [ai]
      summary: Run full FMART analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeConceptBody"
      responses:
        "200":
          description: Feasibility report
          content:
            application/json:
              schema:
                type: object
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    CommentItem:
      type: object
      properties:
        id:
          type: string
        reportId:
          type: string
        userId:
          type: string
        section:
          type: string
          nullable: true
        body:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reportId
        - userId
        - body
        - createdAt
    CommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentItem"
        profiles:
          type: object
      required:
        - comments
        - profiles
    AddCommentBody:
      type: object
      properties:
        body:
          type: string
        section:
          type: string
          nullable: true
      required:
        - body
...
    AutofillBriefBody:
      type: object
      properties:
        brief:
          type: string
      required:
        - brief
    CompleteFieldBody:
      type: object
      properties:
        field:
          type: string
        partial:
          type: string
        inputs:
          type: object
      required:
        - field
        - partial
        - inputs
    AnalyzeConceptBody:
      type: object
      properties:
        inputs:
          type: object
      required:
        - inputs