# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List user's reports
 */
export const ListReportsResponseItem = zod.object({
  id: zod.string(),
  slug: zod.string(),
  title: zod.string(),
  industry: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListReportsResponse = zod.array(ListReportsResponseItem);

/**
 * @summary Save a new report
 */
export const CreateReportBody = zod.object({
  title: zod.string(),
  industry: zod.string().nullish(),
  inputs: zod.object({}).passthrough(),
  output: zod.object({}).passthrough(),
});

/**
 * @summary Get a report by ID
 */
export const GetReportParams = zod.object({
  id: zod.coerce.string(),
});

export const GetReportResponse = zod.object({
  id: zod.string(),
  slug: zod.string(),
  userId: zod.string(),
  title: zod.string(),
  industry: zod.string().nullish(),
  inputs: zod.object({}).passthrough(),
  output: zod.object({}).passthrough(),
  status: zod.string(),
  isPublic: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a report
 */
export const DeleteReportParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Update report status
 */
export const UpdateReportStatusParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateReportStatusBody = zod.object({
  status: zod.string(),
  note: zod.string().nullish(),
});

export const UpdateReportStatusResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get a report by slug (public)
 */
export const GetReportBySlugParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetReportBySlugResponse = zod.object({
  id: zod.string(),
  slug: zod.string(),
  userId: zod.string(),
  title: zod.string(),
  industry: zod.string().nullish(),
  inputs: zod.object({}).passthrough(),
  output: zod.object({}).passthrough(),
  status: zod.string(),
  isPublic: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get report output only
 */
export const GetReportOutputParams = zod.object({
  id: zod.coerce.string(),
});

export const GetReportOutputResponse = zod.object({}).passthrough();

/**
 * @summary List user notifications
 */
export const ListNotificationsResponseItem = zod.object({
  id: zod.string(),
  kind: zod.string(),
  title: zod.string(),
  body: zod.string().nullish(),
  url: zod.string().nullish(),
  readAt: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListNotificationsResponse = zod.array(
  ListNotificationsResponseItem,
);

/**
 * @summary Mark all notifications as read
 */
export const MarkAllNotificationsReadResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Mark a notification as read
 */
export const MarkNotificationReadParams = zod.object({
  id: zod.coerce.string(),
});

export const MarkNotificationReadResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Delete a notification
 */
export const DeleteNotificationParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List comments for a report
 */
export const ListCommentsParams = zod.object({
  reportId: zod.coerce.string(),
});

export const ListCommentsQueryParams = zod.object({
  section: zod.coerce.string().optional(),
});

export const ListCommentsResponse = zod.object({
  comments: zod.array(
    zod.object({
      id: zod.string(),
      reportId: zod.string(),
      userId: zod.string(),
      section: zod.string().nullish(),
      body: zod.string(),
      createdAt: zod.string(),
    }),
  ),
  profiles: zod.object({}).passthrough(),
});

/**
 * @summary Add a comment
 */
export const AddCommentParams = zod.object({
  reportId: zod.coerce.string(),
});

export const AddCommentBody = zod.object({
  body: zod.string(),
  section: zod.string().nullish(),
});

/**
 * @summary Get current user profile
 */
export const GetMyProfileResponse = zod.object({
  userId: zod.string(),
  displayName: zod.string().nullish(),
  avatarUrl: zod.string().nullish(),
});

/**
 * @summary Auto-fill form from a brief
 */
export const AutofillBriefBody = zod.object({
  brief: zod.string(),
});

export const AutofillBriefResponse = zod.object({}).passthrough();

/**
 * @summary AI complete a field
 */
export const CompleteFieldBody = zod.object({
  field: zod.string(),
  partial: zod.string(),
  inputs: zod.object({}).passthrough(),
});

export const CompleteFieldResponse = zod.object({}).passthrough();

/**
 * @summary Run full FMART analysis
 */
export const AnalyzeConceptBody = zod.object({
  inputs: zod.object({}).passthrough(),
});

export const AnalyzeConceptResponse = zod.object({}).passthrough();