# lib/api-zod/src/generated/types/commentItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommentItem {
  id: string;
  reportId: string;
  userId: string;
  section?: string | null;
  body: string;
  createdAt: string;
}