# lib/api-zod/src/generated/types/commentsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommentItem } from "./commentItem";
import type { CommentsResponseProfiles } from "./commentsResponseProfiles";

export interface CommentsResponse {
  comments: CommentItem[];
  profiles: CommentsResponseProfiles;
}