# lib/api-zod/src/generated/types/completeFieldBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompleteFieldBodyInputs } from "./completeFieldBodyInputs";

export interface CompleteFieldBody {
  field: string;
  partial: string;
  inputs: CompleteFieldBodyInputs;
}